+++ CHANGES.md
@@ -25,6 +25,10 @@ development source code and as such may not be routinely kept up to date.
   considered safe/suitable for use.
   ([#416](https://github.com/nextstrain/cli/pull/416))
 
+* `nextstrain view` now correctly handles IPv6 host addresses in the URL opened
+  in a browser (if any).
+  ([#417](https://github.com/nextstrain/cli/pull/417))
+
 
 # 8.5.4 (1 November 2024)